realign buffer modules
